+++ scheduler/job.c
@@ -306,8 +306,8 @@ cupsdCheckJobs(void)
 	  continue;
       }
 
-      cupsdSetJobState(job, IPP_JOB_PENDING, CUPSD_JOB_DEFAULT,
-                       "Job submission timed out.");
+      cupsdSetJobState(job, IPP_JOB_PENDING, CUPSD_JOB_DEFAULT, "Job submission timed out.");
+      cupsdLogJob(job, CUPSD_LOG_ERROR, "Job submission timed out.");
     }
 
    /*