+++ CHANGES.md
@@ -5,6 +5,7 @@ CHANGES - 2.2.12 - 2019-04-15
 Changes in CUPS v2.2.12
 -----------------------
 
+- Updated the systemd service file for cupsd (Issue #5551)
 - The scheduler did not encode octetString values like "job-password" correctly
   for the print filters (Issue #5558)