+++ CHANGES.md
@@ -1,4 +1,4 @@
-CHANGES - 2.3b8 - 2019-04-15
+CHANGES - 2.3b8 - 2019-04-18
 ============================
 
 
@@ -9,6 +9,8 @@ Changes in CUPS v2.3b8
 - The lpadmin command would hang with a bad PPD file (rdar://41495016)
 - Fixed a potential crash bug in cups-driverd (rdar://46625579)
 - Fixed a performance regression with large PPDs (rdar://47040759)
+- Fixed a memory reallocation bug in HTTP header value expansion
+  (rdar://problem/50000749)
 - Restored minimal support for the `Emulators` keyword in PPD files to allow
   old Samsung printer drivers to continue to work (Issue #5562)
 - The scheduler did not encode octetString values like "job-password" correctly